Lights with Screen On [1/2]: Optional allowment of lights
Implement a setting allowing lights to be activated
for new notifications even if the screen is on.

Lights with screen on and Custom values are separated
in an advanced section for a cleaner overview.

This setting gives the user an oportunity to activate lights
with the screen on and also during DayDream screensaver.
The option is not activated by default.

Changes include :
  frameworks/base
  packages/Apps/Settings

Screenshot of the Settings :
  http://i1285.photobucket.com/albums/a583/adriandc/Screenshot_2015-07-20-23-47-13_zpstpmemwwn.png~original

Change-Id: I2071147d8ddab80ba0e1e7310e785ac3e03b3c7c
Signed-off-by: AdrianDC <radian.dc@gmail.com>

Lights with screen on: Don't disable leds after the lockscreen

- If lights with screen on is enabled, do not disable leds
  when user passes through the lockscreen.

Signed-off-by: spezi77 <spezi77@gmx.com>