sensors: Create bool to select what timestamp to use
Older devices may have an issue with rotation freezes up and
requires a reboot to fix. In deep sleep the sensor's timestamp
is far off, depending how long it's in sleep, causing rotation
not to work. onSensorChanged if true it will use
SystemClock.elapsedRealtimeNanos() instead of event.timestamp.
Possibly an update to the custom sensor libs.

Change-Id: Ie456e12cb65fbb921cb780112df301655b93b14f

sensors: Rename bool for selecting what timestamp to use for orientation sensor

Change-Id: I4e91c48ad1520a8a88a92e18ab7e71edaf313ba2
Signed-off-by: spezi77 <spezi77@gmx.com>