Fix memory leak in WifP2pManager
Per WifiP2pManager has an AyncChannel instance. Due to commit
19c4883c850c2d1628e844b48778856673d8bfe6, it aimed at adding
death monitor to AsyncChannel Server for each AyncChannel
Client. But currently, there is no way to unlink the death
monitor unless server is dead. This caused AyncChannel instance
of WifiP2pManager cannot be GCed and resulted in memory leak.

This commit modifies that mechanism, allowing only one channel
instance to exist per process, so the previous existing channels
will be disconnected. But if application implements the callback of
ChannelListener.onChannelDisconnected(), AyncChannel's disconnection
could be noticed by it and this impact can be avoided.

Bug: 37443149
Test: Manual test

Change-Id: I769e4ac990eeff49e10c09a32fba89c7314c21f3