Fix Android Services Library crash
[Cause of Defect]
NotificationManagerService listens to many broadcast and remove received
packages and restart the remote service again.
Remote service would ipc call NotificationManagerService and the ipc
call would do some check.
The whole procedure is finished with many messages and different thread.
When some broadcast comes to NotificationManagerService in a special
moment, the check fail and client crashes.
The detailed analysis is put in the issue:
https://issuetracker.google.com/issues/63554270

Bug: https://issuetracker.google.com/issues/63554270
Test: manual - check https://github.com/nanjingdaqi/issue_demo

Change-Id: I035155d564674d1d71ef7413af1abccbbc126473
Signed-off-by: daqi <daqi@xiaomi.com>
Signed-off-by: mydongistiny <jaysonedson@gmail.com>
Signed-off-by: spezi77 <spezi77@gmx.com>